refactor: move `playground` section to its own page to add the demo
